requirements.txt
Django==1.8.6
django-uuslug==1.1.8
olefile==0.44
Pillow==4.1.1
python-slugify==1.2.4
pytz==2017.2
redis==2.10.5
sorl-thumbnail==12.3
Unidecode==0.4.20
